test: expand rbac action tests
